requirements.txt
streamlit
scikit-learn
seaborn
matplotlib
faker
kiwi
kiwipiepy
python-pptx
python-docx
streamlit-cropper
pillow
pdf2image